I have done everything I can.
I am stuck since I cant find a way to add the first two patients into the double linked list since there are various nulls that intertwin with each other when there is only 0-1 nodes. This would require a lot more if statements and methods and I don't have that much time. Will just finish commenting the code and upload it. Might ask a TA tomorrow or Tuesday to see how I was "Meant to do it"